moved endpoints, added template saving option moved terms and templates, as well as user into their own endpoint; templates are now gathered from swate alpha online; templates edited and saved from the template editor are saved locally on the backend
