+++ app/api/endpoints/projects.py
@@ -41,8 +41,6 @@ from app.api.IO.excelIO import (
     readIsaFile,
     getIsaType,
     writeIsaFile,
-    createSheet,
-    getSwateSheets,
     appendAssay,
     appendStudy,
 )
@@ -52,19 +50,13 @@ from app.models.gitlab.input import (
     folderContent,
     newIsa,
     isaContent,
-    sheetContent,
     syncAssayContent,
     syncStudyContent,
-    userContent,
 )
 from app.models.gitlab.projects import Projects
 from app.models.gitlab.arc import Arc
 from app.models.gitlab.commit import Commit
 from app.models.gitlab.file import FileContent
-from app.models.gitlab.user import Users
-from app.models.swate.template import Templates
-from app.models.swate.templateBuildingBlock import TemplateBB
-from app.models.swate.term import Terms
 
 import hashlib
 import tempfile
@@ -1530,441 +1522,6 @@ async def uploadFile(
 
 
 @router.get(
-    "/getTemplates",
-    summary="Retrieve a list of swate templates",
-    status_code=status.HTTP_200_OK,
-)
-async def getTemplates() -> Templates:
-    startTime = time.time()
-    # send get request to swate api requesting all templates
-    request = requests.get(
-        "https://swate-alpha.nfdi4plants.org/api/ITemplateAPIv1/getTemplates"
-    )
-    try:
-        templateJson = request.json()
-        templateList = []
-        for x in templateJson:
-            templateList.append(json.loads(x))
-    except:
-        templateJson = {}
-
-    # if swate is down, return error 500
-    if not request.ok:
-        logging.error(
-            f"There was an error retrieving the swate templates! ERROR: {templateJson}"
-        )
-        writeLogJson(
-            "getTemplates",
-            500,
-            startTime,
-            f"There was an error retrieving the swate templates! ERROR: {templateJson}",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Couldn't receive swate templates",
-        )
-
-    # map the received list to the model 'Templates'
-    template_list = Templates(templates=templateList)
-
-    logging.info("Sent list of swate templates to client!")
-    writeLogJson(
-        "getTemplates",
-        200,
-        startTime,
-    )
-    # return the templates
-    return template_list
-
-
-@router.get(
-    "/getTemplate",
-    summary="Retrieve the specific template",
-    status_code=status.HTTP_200_OK,
-)
-async def getTemplate(id: str) -> TemplateBB:
-    startTime = time.time()
-    # wrap the desired id in an json array
-    payload = json.dumps([id])
-
-    logging.debug("Getting template with id: " + id)
-    # request the template
-    request = requests.post(
-        "https://swate.nfdi4plants.org/api/IProtocolAPIv1/getProtocolById",
-        data=payload,
-    )
-    try:
-        templateJson = request.json()
-    except:
-        templateJson = {}
-    # if swate is down (or the desired template somehow not available) return error 400
-    if not request.ok:
-        logging.error(
-            f"There was an error retrieving the swate template with id {id} ! ERROR: {templateJson}"
-        )
-        writeLogJson(
-            "getTemplate",
-            400,
-            startTime,
-            f"There was an error retrieving the swate template with id {id} ! ERROR: {templateJson}",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Couldn't find template with id: " + id,
-        )
-
-    # return just the buildingBlocks part of the template (rest is already known)
-    templateBlocks = TemplateBB(templateBB=templateJson["TemplateBuildingBlocks"])
-
-    logging.info(f"Sending template with id {id} to client!")
-    writeLogJson(
-        "getTemplate",
-        200,
-        startTime,
-    )
-    return templateBlocks
-
-
-@router.get(
-    "/getTerms",
-    summary="Retrieve Terms for the given query and parent term",
-    status_code=status.HTTP_200_OK,
-)
-async def getTerms(
-    input: str,
-    parentName: str,
-    parentTermAccession: str,
-    advanced=False,
-) -> Terms:
-    startTime = time.time()
-    # the following requests will timeout after 7s (10s for extended), because swate could otherwise freeze the backend by not returning any answer
-    try:
-        # if there is an extended search requested, make an advanced search call
-        if advanced == "true":
-            request = requests.post(
-                "https://swate.nfdi4plants.org/api/IOntologyAPIv2/getTermsForAdvancedSearch",
-                data=json.dumps(
-                    [
-                        {
-                            "Ontologies": None,
-                            "TermName": input,
-                            "TermDefinition": "",
-                            "KeepObsolete": False,
-                        }
-                    ]
-                ),
-                timeout=10,
-            )
-            logging.debug(f"Getting an extended list of terms for the input '{input}'!")
-        else:
-            # default is an request call containing the parentTerm values
-            request = requests.post(
-                "https://swate.nfdi4plants.org/api/IOntologyAPIv2/getTermSuggestionsByParentTerm",
-                data=json.dumps(
-                    [
-                        {
-                            "n": 20,
-                            "parent_term": {
-                                "Name": parentName,
-                                "TermAccession": parentTermAccession,
-                            },
-                            "query": input,
-                        }
-                    ]
-                ),
-                timeout=7,
-            )
-            logging.debug(
-                f"Getting an specific list of terms for the input '{input}' with parent '{parentName}'!"
-            )
-        try:
-            termJson = request.json()
-        except:
-            termJson = {}
-    # if there is a timeout, respond with an error 504
-    except requests.exceptions.Timeout:
-        logging.warning("Request took to long! Sending timeout error to client...")
-        writeLogJson(
-            "getTerms",
-            504,
-            startTime,
-            "Request took to long! Sending timeout error to client...",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_504_GATEWAY_TIMEOUT,
-            detail="No term could be found in time!",
-        )
-
-    # if there is a different kind of error, return error 400
-    except:
-        logging.error(
-            f"There was an error retrieving the terms for '{input}'! ERROR: {termJson}"
-        )
-        writeLogJson(
-            "getTerms",
-            400,
-            startTime,
-            f"There was an error retrieving the terms for '{input}'! ERROR: {termJson}",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Your request couldn't be processed!",
-        )
-
-    logging.info(f"Sent a list of terms for '{input}' to client!")
-    writeLogJson("getTerms", 200, startTime)
-    try:
-        output = Terms(terms=termJson)
-    except:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="No valid Terms could be found/parsed!",
-        )
-
-    # return the list of terms found for the given input
-    return output
-
-
-@router.get(
-    "/getTermSuggestionsByParentTerm",
-    summary="Retrieve Term suggestions for the given parent term",
-    status_code=status.HTTP_200_OK,
-)
-async def getTermSuggestionsByParentTerm(
-    parentName: str, parentTermAccession: str
-) -> Terms:
-    startTime = time.time()
-    # the following requests will timeout after 7s (10s for extended), because swate could otherwise freeze the backend by not returning any answer
-    try:
-        # default is an request call containing the parentTerm values
-        request = requests.post(
-            "https://swate.nfdi4plants.org/api/IOntologyAPIv2/getAllTermsByParentTerm",
-            data=json.dumps(
-                [
-                    {
-                        "Name": parentName,
-                        "TermAccession": parentTermAccession,
-                    }
-                ]
-            ),
-            timeout=7,
-        )
-        logging.debug(
-            f"Getting list of suggestion terms for the parent '{parentName}'!"
-        )
-        try:
-            termJson = request.json()
-        except:
-            termJson = {}
-    # if there is a timeout, respond with error 504
-    except requests.exceptions.Timeout:
-        logging.warning("Request took to long! Sending timeout error to client...")
-        writeLogJson(
-            "getTermSbPT",
-            504,
-            startTime,
-            "Request took to long! Sending timeout error to client...",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_504_GATEWAY_TIMEOUT,
-            detail="No terms could be found in time!",
-        )
-
-    # if there is a different kind of error, return error 400
-    except:
-        logging.error(
-            f"There was an error retrieving the terms for '{parentName}'! ERROR: {termJson}"
-        )
-        writeLogJson(
-            "getTermSbPT",
-            400,
-            startTime,
-            f"There was an error retrieving the terms for '{parentName}'! ERROR: {termJson}",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Your request couldn't be processed!",
-        )
-    logging.info(f"Sent a list of terms for '{parentName}' to client!")
-    writeLogJson(
-        "getTermSbPT",
-        200,
-        startTime,
-    )
-    try:
-        output = Terms(terms=termJson)
-    except:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="No valid Terms could be found/parsed!",
-        )
-
-    # return the list of terms found for the given input
-    return output
-
-
-@router.get(
-    "/getTermSuggestions",
-    summary="Retrieve Term suggestions by given input",
-    status_code=status.HTTP_200_OK,
-)
-async def getTermSuggestions(input: str, n=20) -> Terms:
-    startTime = time.time()
-    # the following requests will timeout after 7s (10s for extended), because swate could otherwise freeze the backend by not returning any answer
-    try:
-        # default is an request call containing the parentTerm values
-        request = requests.post(
-            "https://swate.nfdi4plants.org/api/IOntologyAPIv2/getTermSuggestions",
-            data=json.dumps(
-                [
-                    {
-                        "n": n,
-                        "ontology": None,
-                        "query": input,
-                    }
-                ]
-            ),
-            timeout=7,
-        )
-        try:
-            termJson = request.json()
-        except:
-            termJson = {}
-
-        logging.debug(f"Getting list of suggestion terms for the input '{input}'!")
-    # if there is a timeout, respond with an error 504
-    except requests.exceptions.Timeout:
-        logging.warning("Request took to long! Sending timeout error to client...")
-        writeLogJson(
-            "getTermSugg.",
-            504,
-            startTime,
-            "Request took to long! Sending timeout error to client...",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_504_GATEWAY_TIMEOUT,
-            detail="No terms could be found in time!",
-        )
-
-    # if there is a different kind of error, return error 400
-    except:
-        logging.error(
-            f"There was an error retrieving the terms for '{input}'! ERROR: {termJson}"
-        )
-        writeLogJson(
-            "getTermSugg.",
-            400,
-            startTime,
-            f"There was an error retrieving the terms for '{input}'! ERROR: {termJson}",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Your request couldn't be processed!",
-        )
-    logging.info(f"Sent a list of terms for '{input}' to client!")
-    writeLogJson(
-        "getTermSugg.",
-        200,
-        startTime,
-    )
-    try:
-        output = Terms(terms=termJson)
-    except:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="No valid Terms could be found/parsed!",
-        )
-
-    # return the list of terms found for the given input
-    return output
-
-
-@router.put(
-    "/saveSheet",
-    summary="Update or save changes to a sheet",
-    status_code=status.HTTP_200_OK,
-)
-async def saveSheet(
-    request: Request, content: sheetContent, data: Annotated[str, Cookie()]
-):
-    startTime = time.time()
-    try:
-        token = getData(data)
-        target = token["target"]
-
-        path = content.path
-        projectId = content.id
-        name = content.name
-
-    # if there are either the name or the accession missing, return error 400
-    except:
-        logging.warning("Client request couldn't be processed, the content is missing!")
-        writeLogJson(
-            "saveSheet",
-            400,
-            startTime,
-            "Client request couldn't be processed, the content is missing!",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail="Couldn't retrieve content of table",
-        )
-
-    # get the file in the backend
-    await arc_file(projectId, path, request, data)
-
-    pathName = f"{os.environ.get('BACKEND_SAVE')}{target}-{projectId}/{path}"
-
-    # if no sheet name is given, name it "sheet1"
-    if name == "":
-        name = "sheet1"
-
-    # add the new sheet to the file
-    createSheet(content, target)
-
-    # send the edited file back to gitlab
-    response = await commitFile(request, projectId, path, data, pathName, message=name)
-    writeLogJson("saveSheet", 200, startTime)
-    return str(response)
-
-
-@router.get(
-    "/getSheets",
-    summary="Get the different annotation metadata sheets of an isa file",
-    status_code=status.HTTP_200_OK,
-)
-async def getSheets(
-    request: Request, path: str, id, data: Annotated[str, Cookie()], branch="main"
-) -> tuple[list, list[str]]:
-    startTime = time.time()
-    try:
-        token = getData(data)
-    except:
-        logging.warning(f"No authorized Cookie found! Cookies: {request.cookies}")
-        writeLogJson(
-            "getSheets",
-            401,
-            startTime,
-            f"No authorized Cookie found!",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="No authorized cookie found!",
-        )
-
-    # get the file in the backend
-    await arc_file(id, path, request, data, branch)
-
-    # construct path to the backend
-    pathName = f"{os.environ.get('BACKEND_SAVE')}{token['target']}-{id}/{path}"
-
-    # read out the list of swate sheets
-    sheets = getSwateSheets(pathName, getIsaType(path))
-    writeLogJson("getSheets", 200, startTime)
-    return sheets
-
-
-@router.get(
     "/getChanges",
     summary="Get the commit history of the ARC",
     status_code=status.HTTP_200_OK,
@@ -2471,298 +2028,6 @@ async def createFolder(
     logging.info(f"Created folder on path: {path}")
     writeLogJson("createFolder", 201, startTime)
     return request.content
-
-
-# get a list of all users for the datahub
-@router.get("/getUser", summary="Get a list of all users")
-async def getUser(request: Request, data: Annotated[str, Cookie()]) -> Users:
-    startTime = time.time()
-    try:
-        token = getData(data)
-        header = {"Authorization": "Bearer " + token["gitlab"]}
-        target = getTarget(token["target"])
-    except:
-        logging.warning(f"No authorized Cookie found! Cookies: {request.cookies}")
-        writeLogJson(
-            "getUser",
-            401,
-            startTime,
-            f"No authorized Cookie found!",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="No authorized cookie found!",
-        )
-
-    userList = []
-    users = requests.head(
-        f"{os.environ.get(target)}/api/v4/users?per_page=100",
-        headers=header,
-    )
-    try:
-        pages = users.headers["x-total-pages"]
-    except:
-        writeLogJson(
-            "getUser",
-            404,
-            startTime,
-            "No users found! Reason: " + users.reason,
-        )
-        raise HTTPException(
-            status_code=users.status_code,
-            detail="No users found! Reason: " + users.reason,
-        )
-
-    for x in range(int(pages)):
-        users = requests.get(
-            f"{os.environ.get(target)}/api/v4/users?per_page=100&without_project_bots=true&page="
-            + str(x + 1),
-            headers=header,
-        )
-        try:
-            userList += users.json()
-        except:
-            userList += {}
-
-    logging.info(f"Sent list of all users of the datahub!")
-    writeLogJson("getUser", 200, startTime)
-
-    output = Users(users=userList)
-
-    return output
-
-
-@router.post(
-    "/addUser",
-    summary="Adds a user to the project",
-    status_code=status.HTTP_201_CREATED,
-)
-async def addUser(
-    request: Request, userData: userContent, data: Annotated[str, Cookie()]
-):
-    startTime = time.time()
-    try:
-        token = getData(data)
-        header = {
-            "Authorization": "Bearer " + token["gitlab"],
-            "Content-Type": "application/x-www-form-urlencoded",
-        }
-        target = getTarget(token["target"])
-    except:
-        logging.warning(f"No authorized Cookie found! Cookies: {request.cookies}")
-        writeLogJson(
-            "addUser",
-            401,
-            startTime,
-            f"No authorized Cookie found!",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="No authorized cookie found!",
-        )
-
-    # get the id and name of the user
-    arcId = userData.id
-    name = userData.username
-    userId = userData.userId
-
-    # look if the user role is set, else set it to 30 (developer)
-    try:
-        userRole = userData.role
-    except:
-        userRole = 30
-
-    addRequest = requests.post(
-        f"{os.environ.get(target)}/api/v4/projects/{arcId}/members",
-        headers=header,
-        data=f"user_id={userId}&access_level={userRole}",
-    )
-    if not addRequest.ok:
-        logging.error(f"Couldn't add user {name} ! ERROR: {addRequest.content}")
-        writeLogJson(
-            "addUser",
-            400,
-            startTime,
-            f"Couldn't add user {name} ! ERROR: {addRequest.content}",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"Couldn't add user to project! Error: {addRequest.content}",
-        )
-    logging.info(f"Added user {name} to project {arcId} with role {userRole}")
-    writeLogJson("addUser", 201, startTime)
-
-    return f"The user {name} was added successfully!"
-
-
-# get a list of all users for the specific Arc
-@router.get("/getArcUser", summary="Get a list of all members of the arc")
-async def getArcUser(
-    request: Request, id: int, data: Annotated[str, Cookie()]
-) -> Users:
-    startTime = time.time()
-    try:
-        token = getData(data)
-        header = {"Authorization": "Bearer " + token["gitlab"]}
-        target = getTarget(token["target"])
-    except:
-        logging.warning(f"No authorized Cookie found! Cookies: {request.cookies}")
-        writeLogJson(
-            "getArcUser",
-            401,
-            startTime,
-            f"No authorized Cookie found!",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="No authorized cookie found!",
-        )
-
-    # request to get all users for the specific arc
-    users = requests.get(
-        f"{os.environ.get(target)}/api/v4/projects/{id}/members?per_page=100",
-        headers=header,
-    )
-    if not users.ok:
-        try:
-            userJson = users.json()
-            userJson["error"] != None
-        except:
-            userJson = {"error": "Not found!", "error_description": "No user found!"}
-        writeLogJson(
-            "getArcUser",
-            users.status_code,
-            startTime,
-            f"{userJson['error']}, {userJson['error_description']}",
-        )
-        raise HTTPException(
-            status_code=users.status_code,
-            detail=f"{userJson['error']}, {userJson['error_description']}",
-        )
-
-    logging.info(f"Sent list of users for project {id}")
-    writeLogJson("getArcUser", 200, startTime)
-    try:
-        userList = users.json()
-    except:
-        userList = users.content
-
-    output = Users(users=userList)
-
-    return output
-
-
-# removes a user from the specific Arc
-@router.delete(
-    "/removeUser",
-    summary="Removes a user from the project",
-)
-async def removeUser(
-    request: Request,
-    id: int,
-    userId: int,
-    username: str,
-    data: Annotated[str, Cookie()],
-):
-    startTime = time.time()
-    try:
-        token = getData(data)
-        header = {
-            "Authorization": "Bearer " + token["gitlab"],
-        }
-        target = getTarget(token["target"])
-    except:
-        logging.warning(f"No authorized Cookie found! Cookies: {request.cookies}")
-        writeLogJson(
-            "removeUser",
-            401,
-            startTime,
-            f"No authorized Cookie found!",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="No authorized cookie found!",
-        )
-
-    # request to delete the given user
-    removeRequest = requests.delete(
-        f"{os.environ.get(target)}/api/v4/projects/{id}/members/{userId}",
-        headers=header,
-    )
-    if not removeRequest.ok:
-        logging.error(
-            f"Couldn't remove user {username} ! ERROR: {removeRequest.content}"
-        )
-        writeLogJson(
-            "removeUser",
-            400,
-            startTime,
-            f"Couldn't remove user {username} ! ERROR: {removeRequest.content}",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"Couldn't remove user from project! Error: {removeRequest.content}",
-        )
-    logging.info(f"Removed user {username} from project {id}")
-    writeLogJson("removeUser", 200, startTime)
-
-    return f"The user {username} was removed successfully!"
-
-
-# edits the role of a user from the specific Arc
-@router.put(
-    "/editUser",
-    summary="Edits a user of the project",
-)
-async def editUser(
-    request: Request, userData: userContent, data: Annotated[str, Cookie()]
-):
-    startTime = time.time()
-    try:
-        token = getData(data)
-        header = {
-            "Authorization": "Bearer " + token["gitlab"],
-        }
-        target = getTarget(token["target"])
-    except:
-        logging.warning(f"No authorized Cookie found! Cookies: {request.cookies}")
-        writeLogJson(
-            "editUser",
-            401,
-            startTime,
-            f"No authorized Cookie found!",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_401_UNAUTHORIZED,
-            detail="No authorized cookie found!",
-        )
-
-    id = userData.id
-    userId = userData.userId
-    username = userData.username
-    role = userData.role
-
-    # request to update user with given role
-    editRequest = requests.put(
-        f"{os.environ.get(target)}/api/v4/projects/{id}/members/{userId}?access_level={role}",
-        headers=header,
-    )
-    if not editRequest.ok:
-        logging.error(f"Couldn't edit user {username} ! ERROR: {editRequest.content}")
-        writeLogJson(
-            "editUser",
-            400,
-            startTime,
-            f"Couldn't edit user {username} ! ERROR: {editRequest.content}",
-        )
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"Couldn't edit user from project {id} to role {role}! Error: {editRequest.content}",
-        )
-    logging.info(f"Edited user {username} from project {id} to role {role}")
-    writeLogJson("editUser", 200, startTime)
-
-    return f"The user {username} was edited successfully!"
 
 
 # sends back a list of metrics for display
