+++ app/models/gitlab/input.py
@@ -64,3 +64,14 @@ class userContent(BaseModel):
     username: str = Field(examples=["lu98be"])
     id: int = Field(examples=[230])
     role: int = Field(examples=[30])
+
+
+class templateContent(BaseModel):
+    table: list
+    name: str
+    identifier: str
+    description: str
+    organisation: str
+    version: str
+    username: dict
+    tags: list
